patient_registry_service/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
paho-mqtt>=1.6.0
